// Arashyn-FE/src/lib/api/generated/models/grammar-summary-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Arashyn API
 * REST API for Arashyn Platform
 *
 * The version of the OpenAPI document: v1.0
 * Contact: touman876@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { GrammarComponentSummaryResponse } from './grammar-component-summary-response';
// May contain unused imports in some cases
// @ts-ignore
import type { GrammarFilterResponse } from './grammar-filter-response';
// May contain unused imports in some cases
// @ts-ignore
import type { GrammarMeaningSummaryResponse } from './grammar-meaning-summary-response';

export interface GrammarSummaryResponse {
    'id'?: string;
    'title'?: string;
    'components'?: Array<GrammarComponentSummaryResponse>;
    'meanings'?: Array<GrammarMeaningSummaryResponse>;
    'filters'?: Array<GrammarFilterResponse>;
}